Extract out tasks into their own package
