+++ src/main/java/com/duke/bot/PersistentStorage.java
@@ -9,6 +9,10 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import com.duke.bot.task.Deadline;
+import com.duke.bot.task.Event;
+import com.duke.bot.task.Task;
+import com.duke.bot.task.Todo;
 
 public class PersistentStorage {
     private final Path storagePath;
