+++ src/main/java/com/duke/bot/command/DoneCommand.java
@@ -3,7 +3,7 @@ package com.duke.bot.command;
 import java.util.ArrayList;
 import java.util.List;
 import com.duke.bot.DukeException;
-import com.duke.bot.Task;
+import com.duke.bot.task.Task;
 
 public class DoneCommand extends Command {
     private final int doneIndex;
