+++ src/main/java/com/duke/bot/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package com.duke.bot.command;
 
 import java.util.List;
-import com.duke.bot.Task;
+import com.duke.bot.task.Task;
 
 public class ExitCommand extends Command {
     @Override
